+++ resources/js/bootstrap/plugins.js
@@ -49,6 +49,14 @@ const defaultPluginOptions = {
     store: null,
     api: null,
 };
+const defaultComponentOptions = {
+    of: null, // id of registered plugin
+    key: null,
+    type: null, // e.g. attribute
+    datatype: null, // required if type=attribute
+    component: null,
+    componentTag: null,
+};
 const defaultSlotOptions = {
     of: null, // id of registered plugin
     slot: 'tab', // one of 'tab', 'tools' or 'settings'
@@ -226,6 +234,35 @@ export const SpPS = {
             }
         }
         SpPS.api.store.systemStore.registerPluginInSlot(mergedOptions);
+    },
+    registerComponent: (options) => {
+        if(!options.of || !SpPS.data.plugins[options.of]) {
+            throw new Error('This plugin part has no associated plugin or that plugin is not installed!');
+        }
+        if(!options.component) {
+            throw new Error('To register a component you must provide a component!');
+        }
+        const mergedOptions = {
+            ...defaultComponentOptions,
+            ...only(options, Object.keys(defaultComponentOptions)),
+        };
+        if(mergedOptions.type != 'attribute') {
+            if(!mergedOptions.componentTag) {
+                mergedOptions.componentTag = mergedOptions.key;
+            }
+            mergedOptions.componentTag = `sp-plugin-${mergedOptions.componentTag}`;
+            if(!!mergedOptions.component) {
+                if(typeof mergedOptions.component == 'string') {
+                    SpPS.data.app.component(mergedOptions.componentTag, {
+                        template: mergedOptions.component,
+                    });
+                } else {
+                    SpPS.data.app.component(mergedOptions.componentTag, mergedOptions.component);
+                }
+            }
+        } else {
+            SpPS.api.store.systemStore.registerPluginAttribute(mergedOptions);
+        }
     },
     registerPreference: (options) => {
         if(!options.of || !SpPS.data.plugins[options.of]) {